"added field validation"

+++ src/main/java/com/outseer/tms/dto/AccountDto.java
@@ -4,11 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class AccountDto {
     private String userId;
-    private Double currentBalance;
+    private BigDecimal currentBalance;
     private Long transactionCount;
 }
